load player base stats in db, calc and sync entities atkspd/movespeed

+++ java-server/src/main/java/com/shnok/javaserver/db/entity/Npc.java
@@ -52,8 +52,8 @@ public class Npc {
     private int rhand;
     private int lhand;
     private int armor;
-    private float walkspd;
-    private float runspd;
+    private int walkspd;
+    private int runspd;
     @Column(name = "faction_id")
     private String factionId;
     @Column(name = "faction_range")
